hak akses admin dan karyawan

+++ Remaster/admin/databunga.php
@@ -1,4 +1,5 @@
 <?php
+  session_start();
   require 'assets/config.php';
 
   $hasil = mysqli_query ($koneksi, "SELECT * FROM bunga");
@@ -49,7 +50,7 @@
   <meta name="description" content="">
   <meta name="author" content="">
 
-  <title>Data Kategori</title>
+  <title>Data Bunga</title>
 
   <!-- Custom fonts for this template -->
   <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
@@ -76,7 +77,12 @@
         <div class="sidebar-brand-icon rotate-n-15">
           <i class="fas fa-snowflake"></i>
         </div>
-        <div class="sidebar-brand-text mx-3">Admin <br> Nursery Polije</div>
+        <div class="sidebar-brand-text mx-3">
+        <?php if ($_SESSION['id_status']=="01") {
+          echo "Admin";
+        }elseif ($_SESSION['id_status']=="02") {
+          echo "Karyawan";
+        } ?> <br> Nursery Polije</div>
       </a>
 
       <!-- Divider -->
@@ -130,16 +136,24 @@
       </li>
 
       <!-- Divider -->
-      <hr class="sidebar-divider">
+      <?php if ($_SESSION['id_status']=="01") { ?>
+        <hr class="sidebar-divider">
+    <?php } ?>
+      
 
       <!-- Heading -->
       <div class="sidebar-heading">
-        Tambah / Edit
+        <?php if ($_SESSION['id_status']=="01") {
+          echo "Tambah / Edit";
+        }else{
+
+        } ?>
       </div>
 
       <!-- Nav Item - Tambah / Edit Bunga Collapse Menu -->
       <li class="nav-item">
-        <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsebunga" aria-expanded="true" aria-controls="collapsebunga">
+        <?php if ($_SESSION['id_status']=="01") { ?>
+          <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsebunga" aria-expanded="true" aria-controls="collapsebunga">
           <i class="fas fa-fw fa-snowflake"></i>
           <span>Bunga
           </span>
@@ -156,10 +170,13 @@
             </a>
           </div>
         </div>
+       <?php }else { ?>
+      <?php }?>    
       </li>
 
       <!-- Nav Item - Tambah / Edit Kategori Bunga Collapse Menu -->
       <li class="nav-item">
+        <?php if ($_SESSION['id_status']=="01") { ?>
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsekategori" aria-expanded="true" aria-controls="collapsekategori">
           <i class="fas fa-fw fa-tag"></i>
           <span>Kategori Bunga
@@ -176,12 +193,15 @@
               <span class="text-primary">Tambah Kategori</span>
             </a>
           </div>
-        </div>
+        </div>  
+      <?php }else{ ?>
+     <?php } ?> 
       </li>
 
       <!-- Nav Item - Tambah / Edit Karyawan Collapse Menu -->
       <li class="nav-item">
-        <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsekaryawan" aria-expanded="true" aria-controls="collapsekaryawan">
+        <?php if ($_SESSION['id_status']=="01") { ?>
+           <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapsekaryawan" aria-expanded="true" aria-controls="collapsekaryawan">
           <i class="fas fa-fw fa-user"></i>
           <span>Karyawan
           </span>
@@ -198,6 +218,8 @@
             </a>
           </div>
         </div>
+        <?php }else{ ?>
+      <?php  } ?>
       </li>
 
       <!-- Divider -->
@@ -370,7 +392,12 @@
             <!-- Nav Item - User Information -->
             <li class="nav-item dropdown no-arrow">
               <a class="nav-link dropdown-toggle" href="#" id="userDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                <span class="mr-2 d-none d-lg-inline text-gray-600 small">Valerie Luna</span>
+                <span class="mr-2 d-none d-lg-inline text-gray-600 small">
+                  <?php if ($_SESSION['id_status']=="01") {
+                    echo $_SESSION["nama_user"];
+                  }else{
+                    echo $_SESSION["nama_user"];
+                  }?></span>
                 <img class="img-profile rounded-circle" src="https://source.unsplash.com/QAB-WJcbgJk/60x60">
               </a>
               <!-- Dropdown - User Information -->
@@ -405,8 +432,19 @@
 
         <!-- Page Heading -->
           <div class="d-sm-flex align-items-center justify-content-between mb-4">
-            <h1 class="h3 mb-0 text-gray-800">Selamat Datang, Admin (Nama Admin)</h1>
-            <a class="d-none d-sm-inline-block btn btn-sm btn-primary shadow-sm text-white" data-toggle="modal" data-target="#tambahBunga">Tambah Bunga</a>
+            <h1 class="h3 mb-0 text-gray-800">
+              Selamat Datang
+              <?php if ($_SESSION['id_status']=="01") {
+                echo "Admin";
+              }elseif ($_SESSION['id_status']=="02") {
+                echo "Karyawan";
+              }?>
+            </h1>
+            <?php if ($_SESSION['id_status']=="01") { ?>
+              <a class="d-none d-sm-inline-block btn btn-sm btn-primary shadow-sm text-white" data-toggle="modal" data-target="#tambahBunga">Tambah Bunga</a>   
+          <?php  }else{ ?>
+         <?php } ?>
+            
           </div>
 
         <!-- #############################################################################################
@@ -590,7 +628,11 @@
                       <th>Harga</th>
                       <th>Stok</th>
                       <th>Deskripsi</th>
-                      <th>Tindakan</th>
+                      <?php if ($_SESSION['id_status']=="01") { ?>
+                        <th>Tindakan</th>
+                     <?php }else{ ?>
+                    <?php } ?>
+                      
                     </tr>
                   </thead>
                   <tbody>
@@ -602,12 +644,16 @@
                       <td><?php echo $row["STOK"]?></td>
                       <td><?php echo $row["DESKRIPSI"]?></td>
                       <td>
-                              <button type="button" class="btn btn-primary" style="width: 40px;" data-toggle="modal" data-target="#editBunga">
-                                <i class="fas fa-edit"></i>
-                              </button>
-                              <a class="btn btn-danger" href="hapusbunga.php?id=<?= $row["ID_BUNGA"]; ?>"onclick="return confirm('Anda yakin ingin menghapus data ini ?')" role="button">
-                                <i class="fas fa-trash"></i>
-                              </a> 
+                        <?php if ($_SESSION['id_status']=="01") { ?>
+                            <button type="button" class="btn btn-primary" style="width: 40px;" data-toggle="modal" data-target="#editBunga">
+                            <i class="fas fa-edit"></i>
+                            </button>
+                            <a class="btn btn-danger" href="hapusbunga.php?id=<?= $row["ID_BUNGA"]; ?>"onclick="return confirm('Anda yakin ingin menghapus data ini ?')" role="button">
+                            <i class="fas fa-trash"></i>
+                            </a>
+                       <?php }else{ ?>
+                    <?php  } ?>
+                               
                       </td>
                     </tr>
                     <?php endwhile;?>
